Folders: add write permissions checks in correct places.

+++ Archivist/src/main/java/com/zorroa/archivist/service/AssetServiceImpl.java
@@ -5,9 +5,11 @@ import com.zorroa.archivist.repository.AssetDao;
 import com.zorroa.archivist.sdk.domain.*;
 import com.zorroa.archivist.sdk.service.AssetService;
 import com.zorroa.archivist.sdk.service.MessagingService;
+import com.zorroa.archivist.security.SecurityUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Component;
 
 /**
@@ -64,11 +66,18 @@ public class AssetServiceImpl implements AssetService {
 
     @Override
     public void addToFolder(Asset asset, Folder folder) {
+        if(!folder.getAcl().hasAccess(SecurityUtils.getPermissionIds(), Access.Write)) {
+            throw new AccessDeniedException("You cannot make changes to this folder");
+        }
         assetDao.addToFolder(asset, folder);
+
     }
 
     @Override
     public void removeFromFolder(Asset asset, Folder folder) {
+        if(!folder.getAcl().hasAccess(SecurityUtils.getPermissionIds(), Access.Write)) {
+            throw new AccessDeniedException("You cannot make changes to this folder");
+        }
         assetDao.removeFromFolder(asset, folder);
     }
